+++ python/GafferArnoldUI/ArnoldAttributesUI.py
@@ -249,9 +249,9 @@ Gaffer.Metadata.registerNode(
 			"description",
 			"""
 			The lights that cause this object to cast shadows.
-			Accepts a set expression or a space separated list of
-			lights. Use \"defaultLights\" to refer to all lights that
-			contribute to illumination by default.
+
+			> Caution : This attribute has been superceded and will be removed. Use
+			> StandardAttributes to set the `shadowedLights` attribute instead.
 			""",
 
 			"layout:section", "Visibility",