+++ Changes.md
@@ -4,15 +4,24 @@
 Improvements
 ------------
 
+- StandardAttributes : Added `shadowedLights` attribute to specify shadow linking. This supercedes the Arnold-specific `shadowGroup` attribute on the ArnoldAttributes node.
 - GafferML :
   - Added support for custom ONNX ops via the `GAFFERML_CUSTOM_OPS_LIBRARIES` environment variable. This should contain a comma-separated list of full paths to libraries containing custom ops.
   - Added support for string tensors.
-- RenderMan : Added support for USDLux lights.
+- Cycles : Added support for shadow linking.
+- RenderMan :
+  - Added support for shadow linking.
+  - Added support for USDLux lights.
 
 Fixes
 -----
 
 - RenderManLight : Fixed orientation of color maps on `PxrRectLight`.
+
+API
+---
+
+- TestCase : Added `assertEqualWithAbsError()` method.
 
 1.5.13.0 (relative to 1.5.12.0)
 ========